fix: Fixed product feedback and comment published row in migration

+++ database/migrations/2022_05_11_130888_create_article_comments_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->string('email')->nullable();
             $table->text('text')->nullable();
 
-            $table->bigIncrements('published')->nullable();
+            $table->integer('published')->nullable();
 
             $table->foreignId('article_id')->constrained();
